drf/requirements.txt
asgiref==3.5.2
asttokens==2.0.8
backcall==0.2.0
black==22.8.0
blis==0.7.9
catalogue==2.0.8
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
confection==0.0.3
contourpy==1.0.5
cryptography==38.0.4
cycler==0.11.0
cymem==2.0.7
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
dj-rest-auth==2.2.5
Django==4.1.3
django-allauth==0.51.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drawnow==0.72.5
entrypoints==0.4
executing==1.0.0
filelock==3.8.0
fonttools==4.38.0
idna==3.4
ipykernel==6.15.3
ipython==8.5.0
jedi==0.18.1
Jinja2==3.1.2
jupyter-core==4.11.1
jupyter_client==7.3.5
kiwisolver==1.4.4
langcodes==3.3.0
MarkupSafe==2.1.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
murmurhash==1.0.9
mypy-extensions==0.4.3
nest-asyncio==1.5.5
numpy==1.23.4
oauthlib==3.2.2
packaging==21.3
parso==0.8.3
pathspec==0.10.1
pathy==0.6.2
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
preshed==3.0.8
prompt-toolkit==3.0.31
psutil==5.9.2
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.2
Pygments==2.13.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.6
pywin32==304
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
smart-open==5.2.1
spacy==3.4.2
spacy-legacy==3.0.10
spacy-loggers==1.0.3
sqlparse==0.4.3
srsly==2.4.5
stack-data==0.5.0
thinc==8.1.5
tomli==2.0.1
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
typer==0.4.2
typing_extensions==4.4.0
tzdata==2022.6
urllib3==1.26.12
virtualenv==20.16.5
wasabi==0.10.1
wcwidth==0.2.5
